PlayerFactory & PlayerValidator: Rename classes. Remove duplicated methods and tests
